+++ NGitLab/Impl/BadgeClient.cs
@@ -17,11 +17,11 @@ internal abstract class BadgeClient
 
     public IEnumerable<Badge> All => _api.Get().GetAll<Badge>(_urlPrefix + "/badges");
 
-    public Badge this[int id] => _api.Get().To<Badge>(_urlPrefix + "/badges/" + id.ToStringInvariant());
+    public Badge this[long id] => _api.Get().To<Badge>(_urlPrefix + "/badges/" + id.ToStringInvariant());
 
     public Badge Create(BadgeCreate badge) => _api.Post().With(badge).To<Badge>(_urlPrefix + "/badges");
 
-    public Badge Update(int id, BadgeUpdate badge) => _api.Put().With(badge).To<Badge>(_urlPrefix + "/badges/" + id.ToStringInvariant());
+    public Badge Update(long id, BadgeUpdate badge) => _api.Put().With(badge).To<Badge>(_urlPrefix + "/badges/" + id.ToStringInvariant());
 
-    public void Delete(int id) => _api.Delete().Execute(_urlPrefix + "/badges/" + id.ToStringInvariant());
+    public void Delete(long id) => _api.Delete().Execute(_urlPrefix + "/badges/" + id.ToStringInvariant());
 }
